Download correct segment with large ttfb
This PR fixes an issue with downloading segments
with bitrate near or equals to network bandwidth
in case if ttfb (time to first byte) is not 0.

```
[       |-----------downloading segment-------------]
^       ^                                           ^
A       B                                           C
```

Legend:
A->B - browser does not download the segment. It waits for
the headers and segment's body.
B->C - browser download the segment's body.

Example:
Bandwidth = 1000kbps
Segment's size = 1000kb
Segment's duration = 1s
TTFB = 0.1s

If player estimates network's bandwidth as 1000kbps, it tries
to download the segment with bitrate 1000kbps. And player
expects that segment will be downloaded in 1s. But real
time to load will be `TTFB + size / bandwidth` and it is
`0.1 + 1000/1000 = 1.1`. To plays such segments without
stalled we should download smaller segments which may be
downloaded in 0.9s.